API Add namespaces, move classes and templates around, remove CSS and images
