+++ _config/config.yml
@@ -1,4 +1,7 @@
-LeftAndMain:
+---
+Name: registryconfig
+---
+SilverStripe\Admin\LeftAndMain:
   extra_requirements_css:
     registry/css/RegistryAdmin.css:
       media: screen
