--- changelog.md
@@ -1,27 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-This project adheres to [Semantic Versioning](http://semver.org/).
-
-## [1.0.10]
-
-* Update translations.
-
-## [1.0.9]
-
-* Added userguide documentation
-* Converted to PSR-2
-* Added standard Scrutinizer config
-* Added standard code of conduct
-* Added standard editor config
-* Added standard Travis config
-* Added standard license
-* Added standard git attributes
-
-## [1.0.8]
-
-* Changelog added.
-* Added CWP keyword
-* Fix Postgres issues
-* Update translations
