--- css/RegistryAdmin.css
@@ -1,3 +1,0 @@
-.icon-registryadmin.icon-16 {
-	background: url("../images/storage.png");
-}
