+++ src/RegistryImportFeed.php
@@ -1,4 +1,11 @@
 <?php
+
+namespace SilverStripe\Registry;
+
+use ArrayList;
+use DataObject;
+use RSSFeed;
+
 class RegistryImportFeed
 {
     protected $modelClass;
@@ -17,7 +24,7 @@ class RegistryImportFeed
             $registryPage = DataObject::get_one('RegistryPage', sprintf('"DataClass" = \'%s\'', $this->modelClass));
             if (($registryPage && $registryPage->exists())) {
                 foreach (array_diff(scandir($path), array('.', '..')) as $file) {
-                    $files->push(new RegistryImportFeed_Entry(
+                    $files->push(new RegistryImportFeedEntry(
                         $file,
                         '',
                         filemtime($path . '/' . $file),
@@ -32,58 +39,5 @@ class RegistryImportFeed
             'registry-feed/latest/' . $this->modelClass,
             singleton($this->modelClass)->singular_name() . ' data import history'
         );
-    }
-}
-class RegistryImportFeed_Entry extends ViewableData
-{
-    protected $title, $description, $date, $link;
-
-    public function __construct($title, $description, $date, $link)
-    {
-        $this->title = $title;
-        $this->description = $description;
-        $this->date = $date;
-        $this->link = $link;
-    }
-
-    public static $casting = array(
-        'Date' => 'SS_Datetime'
-    );
-
-    public function Link()
-    {
-        return $this->link;
-    }
-
-    public function Description()
-    {
-        return $this->description;
-    }
-
-    public function Title()
-    {
-        return $this->title;
-    }
-
-    public function Date()
-    {
-        return $this->date;
-    }
-}
-class RegistryImportFeed_Controller extends Controller
-{
-    private static $allowed_actions = array(
-        'latest'
-    );
-
-    public static $url_handlers = array(
-        '$Action/$ModelClass' => 'handleAction',
-    );
-
-    public function latest($request)
-    {
-        $feed = new RegistryImportFeed();
-        $feed->setModelClass($request->param('ModelClass'));
-        return $feed->getLatest()->outputToBrowser();
     }
 }
