+++ tests/Stub/RegistryPageTestPage.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace SilverStripe\Registry\Tests\Stub;
+
+use RegistryPage;
+use TestOnly;
+
+class RegistryPageTestPage extends RegistryPage implements TestOnly
+{
+}
