+++ tests/Stub/RegistryPageTestPageController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace SilverStripe\Registry\Tests\Stub;
+
+use TestOnly;
+use RegistryPageController;
+
+class RegistryPageTestPageController extends RegistryPageController implements TestOnly
+{
+}
