+++ tests/Stub/RegistryPageTestSubclass.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace SilverStripe\Registry\Tests\Stub;
+
+use RegistryPage;
+use TestOnly;
+use RegistryPage_Controller;
+
+class RegistryPageTestSubclass extends RegistryPage implements TestOnly
+{
+}
