+++ tests/fixtures/RegistryPageFunctionalTest.yml
@@ -1,10 +1,10 @@
-RegistryPageTestPage:
-    contact-registrypage:
-        Title: Contact search
-        URLSegment: contact-search
-        DataClass: RegistryPageTestContact
-    contact-registrypage-limit:
-        Title: Contact search
-        URLSegment: contact-search-limit
-        DataClass: RegistryPageTestContact
-        PageLength: 3
+SilverStripe\Registry\RegistryPageTestPage:
+  contact-registrypage:
+    Title: Contact search
+    URLSegment: contact-search
+    DataClass: SilverStripe\Registry\RegistryPageTestContact
+  contact-registrypage-limit:
+    Title: Contact search
+    URLSegment: contact-search-limit
+    DataClass: SilverStripe\Registry\RegistryPageTestContact
+    PageLength: 3
