--- tests/pagetypes/RegistryPageTestPage.php
@@ -1,7 +1,0 @@
-<?php
-class RegistryPageTestPage extends RegistryPage implements TestOnly
-{
-}
-class RegistryPageTestPage_Controller extends RegistryPage_Controller implements TestOnly
-{
-}
