--- tests/pagetypes/RegistryPageTestSubclass.php
@@ -1,7 +1,0 @@
-<?php
-class RegistryPageTestSubclass extends RegistryPage implements TestOnly
-{
-}
-class RegistryPageTestSubclass_Controller extends RegistryPage_Controller implements TestOnly
-{
-}
